AWS/Bedrock_Arize_app/requirements.txt
arize==7.48.1
boto3==1.39.3
pandas==2.3.1
Pillow==11.3.0
python-dotenv==1.1.1
streamlit==1.45.1
streamlit_chat==0.1.1
streamlit_feedback==0.1.4
openinference-instrumentation-bedrock==0.1.26
arize-otel==0.9.0
xlsxwriter==3.2.5
